Rename FPArrayComparator -> FloatingPointComparator
