+++ source/matrix_free/mapping_info.cc
@@ -29,38 +29,4 @@ DEAL_II_NAMESPACE_OPEN
 #endif
 #include "mapping_info.inst"
 
-#if SPLIT_INSTANTIATIONS_INDEX == 0
-
-template struct internal::MatrixFreeFunctions::FPArrayComparator<double>;
-template struct internal::MatrixFreeFunctions::FPArrayComparator<float>;
-
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<double, 1>>;
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<float, 1>>;
-
-#  if (DEAL_II_VECTORIZATION_WIDTH_IN_BITS >= 128 && defined(__SSE2__)) || \
-    (DEAL_II_VECTORIZATION_WIDTH_IN_BITS >= 128 && defined(__ALTIVEC__))
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<double, 2>>;
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<float, 4>>;
-#  endif
-
-#  if DEAL_II_VECTORIZATION_WIDTH_IN_BITS >= 256 && defined(__AVX__)
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<double, 4>>;
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<float, 8>>;
-#  endif
-
-#  if DEAL_II_VECTORIZATION_WIDTH_IN_BITS >= 512 && defined(__AVX512F__)
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<double, 8>>;
-template struct internal::MatrixFreeFunctions::FPArrayComparator<
-  VectorizedArray<float, 16>>;
-#  endif
-
-#endif
-
 DEAL_II_NAMESPACE_CLOSE
